Merge in changes from main repo (#4)
* Player Node  (#47)

Adds a player node which receives data from Vision node for positions and Coach node for strategy. Two source files, one contains robot class and the other updates values for the players and one script to start the player node.

* Shoot skill (#49)

* Started shooting implementation

* Implemented states in the shooting skill

* Finished smart aiming

The shooting robot will now take that shot that results in the least chance that the ball will be blocked.